adding Project Handle Logic

+++ src/pages/Dashboard.js
@@ -1,4 +1,4 @@
-import React, {useState, useRef} from 'react'
+import React, {useState, useRef,useContext} from 'react'
 import { Link, useHistory } from 'react-router-dom'
 import {useAuth} from '../service/AuthContext'
 import {useTheme} from '../service/ThemeContext'
@@ -7,28 +7,42 @@ import { AiFillSetting } from 'react-icons/ai'
 import Logo from '../components/Logo'
 import BottomNav from '../components/BottomNav'
 import TopNav from '../components/TopNav'
+import { AiOutlinePlus } from 'react-icons/ai'
+import { AppTransitionContext } from '../service/AppTransitionContext'
 
 
 export default function Dashboard() {
     const [error, setError] = useState('')
     const {currentUser} = useAuth()
     const { themePrimary, toggleTheme } = useTheme()
-    const { projects, addProject, deleteProject } = useDB()
+    const { projects, deleteProject, addTaskToProject } = useDB()
     const history = useHistory()
+    const { setPreset } = useContext(AppTransitionContext);
     
 
     const nameRef = useRef()
     function changeColor(){
         toggleTheme(themePrimary)
     }
-
-    function createProject(){
-        console.log("Create a new Project")
-        addProject(nameRef.current.value,currentUser.email)
+    async function handleToProjectCreate(){
+        try {
+            await setPreset("cubeToBottom")
+            history.push('/create-project')
+            // TODO set Dashboard & Taskboard seperate
+            // await setPreset("cubeToTop")
+        } catch (error) {
+            console.log(error)
+        }
     }
-    function deleteFromProject(projectid){
+
+    function handleTaskToProject(projectid,task,repeat){
+        addTaskToProject(projectid,task,repeat)
+    }
+    function deleteFromProject(projectid,projectname){
         console.log(`Delete this project ${projectid} now`)
-        deleteProject(projectid)
+        if(window.confirm(`Delete everything on ${projectname}?`)){
+            deleteProject(projectid)
+        }
     }
 
     const styleTest = {
@@ -42,15 +56,14 @@ export default function Dashboard() {
                 <div className="main-container">
                     {error && <h1>{error}</h1>}
                     <div className="projects">
-                        <div className="project">
-                            <input type="text" placeholder="Name" ref={nameRef}/>
-                            <button onClick={createProject} style={styleTest}>Create Project</button>
-                        </div>
+                        <button onClick={handleToProjectCreate} className="project flex-center">
+                            <AiOutlinePlus size="4em"/>
+                        </button>
                         {projects.map((item,i) =>  (
-                        <div key={i} className="project">
+                        <div key={i} className="project" style={{backgroundColor:item.data.color}}>
                             <div>{item.data.name}</div>
-                            <button> Open </button>
-                            <button onClick={()=>deleteFromProject(item.id)}> Delete </button>
+                            <button onClick={()=>handleTaskToProject(item.id,"Geschirrspüler",false)}> Open </button>
+                            <button onClick={()=>deleteFromProject(item.id,item.data.name)}> Delete </button>
                         </div>
                         ))}
                     </div>
